# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warubi Platform - FC Köln ITP API
  description: |
    RESTful API for the 1.FC Köln International Talent Program management system.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Tokens are obtained via the `/api/fckoln/auth/login` endpoint.
    
    ## Multi-Tenancy
    This API supports multiple applications via the `X-App-Key` header.
    Contact support to obtain your application credentials.
    
    ## Rate Limiting
    Requests are rate limited per application. Check response headers for:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
  version: 1.0.0
  contact:
    name: Warubi Platform Support
    email: support@warubi.io
servers:
  - url: /api/fckoln
    description: FC Köln ITP API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Invalid request
        correlationId:
          type: string
          format: uuid
          description: Request correlation ID for debugging
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 50
        total:
          type: integer
          example: 125
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        eventType:
          type: string
          enum: [training, match, meeting, medical, other]
        date:
          type: string
          format: date
        startTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        endTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
        location:
          type: string
        notes:
          type: string
        isRecurring:
          type: boolean
        recurringPattern:
          type: string
          enum: [daily, weekly, monthly]
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
            enum: [training, match, meeting, medical, other]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Events]
      summary: Create event
      description: Staff/Admin only
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, eventType, date, startTime, endTime]
              properties:
                title:
                  type: string
                eventType:
                  type: string
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                endTime:
                  type: string
                location:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Event created
        '403':
          $ref: '#/components/responses/Forbidden'
  /events/{id}:
    put:
      tags: [Events]
      summary: Update event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Events]
      summary: Delete event
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /events/{id}/attendance:
    get:
      tags: [Events]
      summary: Get event attendance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Attendance list
    post:
      tags: [Events]
      summary: Record attendance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                userId:
                  type: integer
                status:
                  type: string
                  enum: [present, absent, late, excused]
      responses:
        '200':
          description: Attendance recorded
...
  /chores:
    get:
      tags: [Chores]
      summary: List chores
      security:
        - BearerAuth: []
      parameters:
        - name: house
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, verified, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Chores list
    post:
      tags: [Chores]
      summary: Create chore
      description: Staff/Admin only
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, house]
              properties:
                title:
                  type: string
                description:
                  type: string
                house:
                  type: string
                assignedTo:
                  type: integer
                dueDate:
                  type: string
                  format: date
      responses:
        '201':
          description: Chore created
        '403':
          $ref: '#/components/responses/Forbidden'
  /chores/{id}/complete:
    post:
      tags: [Chores]
      summary: Mark chore complete
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chore marked complete
  /chores/{id}/verify:
    put:
      tags: [Chores]
      summary: Verify chore completion
      description: Staff/Admin only
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [verified]
              properties:
                verified:
                  type: boolean
      responses:
        '200':
          description: Chore verified/rejected
        '403':
          $ref: '#/components/responses/Forbidden'
...